requirements.txt
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.3.9
chardet==3.0.4
freezegun==0.3.11
idna==2.8
isort==4.3.19
lazy-object-proxy==1.4.1
mccabe==0.6.1
more-itertools==7.0.0
pluggy==0.11.0
py==1.8.0
pylint==2.3.1
pytest==4.5.0
pytest-freezegun==0.3.0.post1
python-dateutil==2.8.0
requests==2.21.0
requests-mock==1.6.0
RPi.bme280==0.2.2
six==1.12.0
smbus2==0.2.3
typed-ast==1.3.5
urllib3==1.24.3
wcwidth==0.1.7
wrapt==1.11.1
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.3.9
chardet==3.0.4
coverage==4.5.3
freezegun==0.3.11
idna==2.8
isort==4.3.19
lazy-object-proxy==1.4.1
mccabe==0.6.1
more-itertools==7.0.0
pluggy==0.11.0
py==1.8.0
pylint==2.3.1
pytest==4.5.0
pytest-cov==2.7.1
pytest-freezegun==0.3.0.post1
python-dateutil==2.8.0
requests==2.21.0
requests-mock==1.6.0
RPi.bme280==0.2.2
six==1.12.0
smbus2==0.2.3
typed-ast==1.3.5
urllib3==1.24.3
wcwidth==0.1.7
wrapt==1.11.1
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.1.0
autopep8==1.4.4
certifi==2019.3.9
chardet==3.0.4
coverage==4.5.3
entrypoints==0.3
flake8==3.7.7
freezegun==0.3.11
idna==2.8
isort==4.3.19
lazy-object-proxy==1.4.1
mccabe==0.6.1
more-itertools==7.0.0
pep8==1.7.1
pluggy==0.11.0
py==1.8.0
pycodestyle==2.5.0
pyflakes==2.1.1
pylint==2.3.1
pytest==4.5.0
pytest-cov==2.7.1
pytest-freezegun==0.3.0.post1
python-dateutil==2.8.0
requests==2.21.0
requests-mock==1.6.0
RPi.bme280==0.2.2
six==1.12.0
smbus2==0.2.3
typed-ast==1.3.5
urllib3==1.24.3
wcwidth==0.1.7
wrapt==1.11.1
